% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivNormality.R
\name{bivNormality}
\alias{bivNormality}
\title{bivNormality}
\usage{
bivNormality(x, alpha = 0.08)
}
\arguments{
\item{x}{the data set contain two variables}
}
\value{
an ellipse around the bivariate data and a list containing
the actual and predicted percentage contents of the ellipse
}
\description{
bivNormality
}
\examples{
bivNormality(x=matrix(c(T4_6$V1, T4_6$V2), byrow=FALSE))
}
