% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxcox.R
\name{boxcox}
\alias{boxcox}
\title{boxcox}
\usage{
boxcox(xs, alpha = 0.05, iter = 1000)
}
\arguments{
\item{xs}{the vector}

\item{alpha}{the ci}

\item{iter}{number of time we are going to run}
}
\description{
carries out the boxcox transformation
}
