% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectingOutliers.R
\name{DetectingOutliers}
\alias{DetectingOutliers}
\title{DetectingOutliers}
\usage{
DetectingOutliers(x)
}
\arguments{
\item{x}{the data set}
}
\value{
the matrix of s, the generalized squared distance, and the outlier
}
\description{
this is the function that would carries out step 1-4 "Setps for Detecting outliers" in textbokk
}
\examples{
DetectingOutliers (T1_4)
}
