% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQQplot.R
\name{getQQplot}
\alias{getQQplot}
\title{return the Q-Q plot of the dataset}
\usage{
getQQplot(data = rnorm(100, mean = 0, sd = 1))
}
\arguments{
\item{data}{the data going use to make Q-Q plot}
}
\value{
the Q-Q plot of the data
}
\description{
return the Q-Q plot of the dataset
}
\examples{
getQQplot(data = rnorm(100, mean = 0, sd = 1))
}
