% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profits.R
\docType{data}
\name{profits}
\alias{profits}
\title{Profits Data}
\format{
A vector with 10 numeric values representing profits figures.
 @source text book
}
\usage{
profits
}
\description{
This data set contains the data of profits in billions
}
\keyword{datasets}
