% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales.R
\docType{data}
\name{sales}
\alias{sales}
\title{Sales Data}
\format{
A vector with 10 numeric values representing sales figures.
}
\source{
text book
}
\usage{
sales
}
\description{
This data set contains data of sales in billions
}
\keyword{datasets}
