% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T4_6.R
\docType{data}
\name{T4_6}
\alias{T4_6}
\title{T1-5 data}
\format{
A data frame with 130 rows and 7 columns
\describe{
  \item{V1}{Indep}
  \item{V2}{Supp}
  \item{V3}{Benev}
  \item{V4}{Conform}
  \item{V5}{Leader}
  \item{V6}{Gender}
  \item{V7}{Socio}
  ...
}
}
\source{
text book
}
\usage{
T4_6
}
\description{
This is a data about psychological test administered to Peruvian
teenagers. This contains 130 measurements with 7 variables
}
\keyword{datasets}
