% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMVN.R
\name{testMVN}
\alias{testMVN}
\title{testMVN}
\usage{
testMVN(x, dim = 2)
}
\arguments{
\item{x}{the dataset, could a a matrix, that going to test}

\item{dim}{the number of variables (columns) in the matrix}
}
\description{
testMVN
}
