% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theoryUnivNormality.R
\name{theoryUnivNormality}
\alias{theoryUnivNormality}
\title{theoryUnivNormality}
\usage{
theoryUnivNormality(x)
}
\arguments{
\item{x}{the dataset going to test univariate Normality through theory}
}
\description{
theoryUnivNormality
}
